Updated codegen templates to reflect latest core sdk changes

+++ workers/unity/Packages/com.improbable.gdk.core/Commands/WorldCommands.cs
@@ -237,7 +237,7 @@ namespace Improbable.Gdk.Core.Commands
 
             internal class Storage : CommandStorage
             {
-                public Dictionary<uint, CommandRequestStore<Request>> CommandRequestsInFlight = new Dictionary<uint, CommandRequestStore<Request>>();
+                public Dictionary<long, CommandRequestStore<Request>> CommandRequestsInFlight = new Dictionary<long, CommandRequestStore<Request>>();
             }
         }
 
@@ -287,7 +287,7 @@ namespace Improbable.Gdk.Core.Commands
 
             internal static class RequestsProvider
             {
-                private static readonly Dictionary<uint, List<Request>> Storage = new Dictionary<uint, List<Request>>();
+                private static readonly Dictionary<long, List<Request>> Storage = new Dictionary<long, List<Request>>();
                 private static readonly Dictionary<uint, World> WorldMapping = new Dictionary<uint, World>();
 
                 private static uint nextHandle = 0;
@@ -417,7 +417,7 @@ namespace Improbable.Gdk.Core.Commands
 
             internal class Storage : CommandStorage
             {
-                public Dictionary<uint, CommandRequestStore<Request>> CommandRequestsInFlight = new Dictionary<uint, CommandRequestStore<Request>>();
+                public Dictionary<long, CommandRequestStore<Request>> CommandRequestsInFlight = new Dictionary<long, CommandRequestStore<Request>>();
             }
         }
 
@@ -467,7 +467,7 @@ namespace Improbable.Gdk.Core.Commands
 
             internal static class RequestsProvider
             {
-                private static readonly Dictionary<uint, List<Request>> Storage = new Dictionary<uint, List<Request>>();
+                private static readonly Dictionary<long, List<Request>> Storage = new Dictionary<long, List<Request>>();
                 private static readonly Dictionary<uint, World> WorldMapping = new Dictionary<uint, World>();
 
                 private static uint nextHandle = 0;
@@ -597,7 +597,7 @@ namespace Improbable.Gdk.Core.Commands
 
             internal class Storage : CommandStorage
             {
-                public Dictionary<uint, CommandRequestStore<Request>> CommandRequestsInFlight = new Dictionary<uint, CommandRequestStore<Request>>();
+                public Dictionary<long, CommandRequestStore<Request>> CommandRequestsInFlight = new Dictionary<long, CommandRequestStore<Request>>();
             }
         }
 
@@ -647,7 +647,7 @@ namespace Improbable.Gdk.Core.Commands
 
             internal static class RequestsProvider
             {
-                private static readonly Dictionary<uint, List<Request>> Storage = new Dictionary<uint, List<Request>>();
+                private static readonly Dictionary<long, List<Request>> Storage = new Dictionary<long, List<Request>>();
                 private static readonly Dictionary<uint, World> WorldMapping = new Dictionary<uint, World>();
 
                 private static uint nextHandle = 0;
@@ -777,7 +777,7 @@ namespace Improbable.Gdk.Core.Commands
 
             internal class Storage : CommandStorage
             {
-                public Dictionary<uint, CommandRequestStore<Request>> CommandRequestsInFlight = new Dictionary<uint, CommandRequestStore<Request>>();
+                public Dictionary<long, CommandRequestStore<Request>> CommandRequestsInFlight = new Dictionary<long, CommandRequestStore<Request>>();
             }
         }
     }
